refactor : #86 Github api 함수 폴더 분리 및 주석 추가
